--- Frontend/src/pages/home/HomeMain.tsx
@@ -1,5 +1,0 @@
-const HomeMain: React.FC = () => {
-  return <></>;
-};
-
-export default HomeMain;
