--- Frontend/src/pages/home/Test.tsx
@@ -1,48 +1,0 @@
-import React, { useEffect, useRef, useState } from 'react';
-import { addMonths, subMonths } from 'date-fns';
-import RenderHeader from './calender/RenderHeader';
-import RenderCells from './calender/RenderCells';
-
-export const HomeCalendar: React.FC = () => {
-  const [currentMonth, setCurrentMonth] = useState(new Date());
-  const [selectedDate, setSelectedDate] = useState<Date>(new Date());
-
-  const prevMonth = () => {
-    setCurrentMonth(subMonths(currentMonth, 1));
-  };
-
-  const nextMonth = () => {
-    setCurrentMonth(addMonths(currentMonth, 1));
-  };
-
-  const onDateClick = (day: Date) => {
-    setSelectedDate(day);
-  };
-
-  const monthRef = useRef<HTMLDivElement>(null);
-
-  useEffect(() => {
-    if (monthRef.current !== null) {
-      monthRef.current.scrollIntoView({ behavior: 'auto' });
-    }
-  }, []);
-
-  return (
-    <div className="schedule-calendar">
-      <div className="calendar__item" ref={monthRef}>
-        <RenderHeader
-          currentMonth={currentMonth}
-          prevMonth={prevMonth}
-          nextMonth={nextMonth}
-        />
-        <RenderCells
-          currentMonth={currentMonth}
-          selectedDate={selectedDate}
-          onDateClick={onDateClick}
-        />
-      </div>
-    </div>
-  );
-};
-
-export default HomeCalendar;
