--- Frontend/src/pages/home/type.ts
@@ -1,11 +1,0 @@
-export interface HomeCalendarProps {
-  currentMonth: Date;
-  prevMonth: () => void;
-  nextMonth: () => void;
-}
-
-export interface RenderCellsProps {
-  currentMonth: Date;
-  selectedDate: Date;
-  onDateClick: (day: Date) => void;
-}
